requirements.txt
streamlit
tensorflow
requests
Pillow
numpy
kaggle